requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
fastapi==0.103.2
greenlet==3.0.3
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==2.5.0
pydantic_core==2.14.1
PyMySQL==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
secure-smtplib==0.1.1
setuptools==68.0.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.22.0
wheel==0.41.2
